----------Angular $http Service---------- In this commit we worked on transforming some of our data stored in services.js into an JSON object format file and allow the Json-Server Node module serve up us those files remotely. To do this we had to change up services.js services object methods (that we use in controllers to call our data) to not refference self, but to have those methods call our json-server via https GET Method. Note after this change everytime we access those data files, we make a GET request to the server. Before changing the services.js file we only had to call the server for the service.js file once.
With error handling we introduced a new directive called ng-if="showMenu" when evaulated to be true will add the DOM element and when not evaluated to true will remove the DOM element competely.

angular $https is a core angular service used to communicate with servers using the HTTP protocol via the browsers request

http operation is ASYNCHRONOUS in nature, results not immediate and time factor has to be accounted for.

in controller code, we have to inisialize our data objects with a blank object, and then we use menuFactory.getDish (call to our object via services.js)