---------------------------Exerice: Angular ngResource module and $resrouce service--------------------------- In this exercise we changed up angular core $http service with an external (needed to be downloaded) ngResource module and $resource service that leverages $htttp protocol to access a RESful API Server.
Web Services: A system designed to support interoperability (abilit of sofware to ex hange and make use of
info) of systems connected of a network. Service Oriented Architecture (SOA, standarizedway of integrating web-
based apps usings open stndards operating over the internet. 2 standards: Soap and REST

REST: WWW was successful b/c hyperlinks aloud to refer to other resources in the internet using url
REST is a style of 'software architecture' for distributed hyperedia systems such as the www
Introduced by Roy Fielding
A collection of network architecture principles which outline how resources are defined and addressed.
Resoureces reffered using hyperlinks
1) itendify resoruecs 2) address resources 3)perform operations on resorouces
4 design principles of REST:
A) use http methods explliicitly
B) be stateless
C) expose directory structure-like URIs
D) JSON
"The motivation for REST was to capture the characteristics of the Web that made the Web Successful"
URI Accesssible Resources

Resources: key abstraction of information in REST, a "conceptual mapping to a set of enteties"
-represented witha global indentifier = URI
Get - Read
Post - Create
Put - Update
Delete - Delete

Representations: JSON, how data is represented or returned to the client. We transform data in a way that can be
transported accroes to the client side or vice versa in a format both sides can udnerstand.
repsresentation: the represented info that is xported across from one side to another is JSON
*** JSON has become the representational standard for information interchange ***

$resource service: higher level of abstraction than the $http service, facilitates communication w/server
that exports RESTful API
-not part of angular core

ngResrouce: convenient wrapper around REST API that a server supports CRUD operations.

DEPENDENCY INJECTION: ngResource injection in app.js modules and $resource injection into services.js
rememeber no $resource injection in controllers.js because controllers.js are injected with our services (menuFactory)

$resource USAGE: ex. $resource( URL. [paramDetails], [actions], options);

$resource default:
.get() - get
.save() - post
.query() - get, isArray: true
.remove() - delete
.delete() - delete

ex. $scope.dishes = $resrouce( baseURL + "dishes/:id", null, {'update' : {method : 'PUT'}} ) .query() ;
"A call to the resrouce will return an object of the resource class, so you can acces the methods immediately.
Also no need to initialize object in controllers, with $resrouce $scope.dishes by defualt will be empty if nothing
is added to it.

Notice above is the dishes object. so we query for the entire array. If we wanted to get particular dish in that object,
we set dish to .getdishes.get with requested id as follows:

$scope.dish = menuFactory.getDishes().get({id:parseInt($stateParams.id,10)
"With this call, I am setting scope.dish to menufac.getdishes, and i use the.get method, with which I supply
parameter as id: and parse the information I get through my stateParams, and pass that as a parameter to .get()
Get method will return the psecific dish I asked for.

in indexcontroller we set scope.dish to .get([id:0})

EXERCISE SUMMARY: Used angular ngResrouce module and $resrouce service to acces a Restful API Server