----ngResource Error Handling----
Had a pretty big bug today. Figured out the issue was my angular and angular resource modules
weren't matching up in versions. Needed to google how to prefix bower install of right version.
We are using angular 1.4.2 i think.
In this exercise we updated our ngResource and $resource to take in errors, similar to how we
error handled with $http.

Also almost was messing this up. Our working copy is different from the distributed copy.